backend/cityconnect/requirements.txt
Django
djangorestframework
django-cors-headers
djangorestframework-simplejwt
Pillow
python-decouple
django-filter